+++ types.hh
@@ -1161,26 +1161,6 @@ typename Type::value_type deserialize_value(Type& t, bytes_view v) {
     return t.deserialize_value(v);
 }
 
-// Does not check bounds. Must be called only after size is already checked.
-template<FragmentedView View>
-void read_fragmented(View& v, size_t n, bytes::value_type* out) {
-    while (n) {
-        if (n <= v.current_fragment().size()) {
-            std::copy_n(v.current_fragment().data(), n, out);
-            v.remove_prefix(n);
-            n = 0;
-        } else {
-            out = std::copy_n(v.current_fragment().data(), v.current_fragment().size(), out);
-            n -= v.current_fragment().size();
-            v.remove_current();
-        }
-    }
-}
-template<> void inline read_fragmented(single_fragmented_view& v, size_t n, bytes::value_type* out) {
-    std::copy_n(v.current_fragment().data(), n, out);
-    v.remove_prefix(n);
-}
-
 template<typename T>
 T read_simple(bytes_view& v) {
     if (v.size() < sizeof(T)) {
@@ -1191,21 +1171,6 @@ T read_simple(bytes_view& v) {
     return net::ntoh(*reinterpret_cast<const net::packed<T>*>(p));
 }
 
-template<typename T, FragmentedView View>
-T read_simple(View& v) {
-    if (v.current_fragment().size() >= sizeof(T)) [[likely]] {
-        auto p = v.current_fragment().data();
-        v.remove_prefix(sizeof(T));
-        return net::ntoh(*reinterpret_cast<const net::packed<T>*>(p));
-    } else if (v.size_bytes() >= sizeof(T)) {
-        T buf;
-        read_fragmented(v, sizeof(T), reinterpret_cast<bytes::value_type*>(&buf));
-        return net::ntoh(buf);
-    } else {
-        throw_with_backtrace<marshal_exception>(format("read_simple - not enough bytes (expected {:d}, got {:d})", sizeof(T), v.size_bytes()));
-    }
-}
-
 template<typename T>
 T read_simple_exactly(bytes_view v) {
     if (v.size() != sizeof(T)) {
@@ -1213,20 +1178,6 @@ T read_simple_exactly(bytes_view v) {
     }
     auto p = v.begin();
     return net::ntoh(*reinterpret_cast<const net::packed<T>*>(p));
-}
-
-template<typename T, FragmentedView View>
-T read_simple_exactly(View v) {
-    if (v.current_fragment().size() == sizeof(T)) [[likely]] {
-        auto p = v.current_fragment().data();
-        return net::ntoh(*reinterpret_cast<const net::packed<T>*>(p));
-    } else if (v.size_bytes() == sizeof(T)) {
-        T buf;
-        read_fragmented(v, sizeof(T), reinterpret_cast<bytes::value_type*>(&buf));
-        return net::ntoh(buf);
-    } else {
-        throw_with_backtrace<marshal_exception>(format("read_simple_exactly - size mismatch (expected {:d}, got {:d})", sizeof(T), v.size_bytes()));
-    }
 }
 
 inline
